Project 2/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alpaca-crypto-lab"
version = "0.1.0"
description = "Crypto-only automated trading and research repository using Alpaca API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.26.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "typer>=0.9.0",
    "loguru>=0.7.0",
    "pyarrow>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"








